feat(link): apply subscription perks mechanism
fix(link): amqp channel could only be consumed once

Previously, the identifier for the `Amqp.consumer` was the channel name.
This means that one channel could only be used by a single consumer
only. Looking at the design (RabbitMQ), I think it was not channel that
better be used once, but rather the queue.

ref(link): make amqp activates monitoring on start

ref(link): send message body to actual consumer

The handler would mostly only need the data, not the technical content
of the message itself